Work around issue that Ctrl+Backspace doesn't trigger its action.
Pressing Ctrl+Backspace never actually "deleted the word to the left",
as it is supposed to. We work around that issue by handling the
low-level key event and distinguishing between Backspace and
Ctrl+Backspace based on the reported "character".

I hope this is robust. I tested in Windows Terminal and Tabby (also on
Windows), both with Windows' CMD shell and using Fish in WSL/Ubuntu.
I'm a bit skeptical because the work-around is so easy, so why wouldn't
Textual just do the same. But it works for all I can tell.