Added key bindings to move cursor to start or end of file.
By default (and currently hard-coded), these are Ctrl+Home and Ctrl+End
respectively. Textual's `TextArea` widget does not define actions for
that, so we added our own. As is always the case, these key bindings
only work if they are not already bound by the terminal application the
editor runs in.