More verbosity in output of `config` command.
Renamed the sub-commands from `dir` and `file` to `folders` and `files`,
as we're showing more than one item now. Avoid "dir" as its an
abbreviation, albeit a common one, and "folder" is shorter than
"directory".

Also, refer to the "global" configuration as the "site" configuration.
Its meaning may be less obvious to the casual user, but it's an
established term nonetheless, and it makes the output look better as the
word is the same length as "user".