Allow key remapping via configuration file.
All application and editor key bindings are now defined via key maps
read from the configuration. The default key bindings are thus defined
in `settings.yaml`, but can be overridden in the user or site config.

Refactored the `config` module to cache the configuration files as we
query each key binding individually, which would otherwise produce a lot
of unnecessary file operations. We invalidate the cache whenever we
store a new configuration value, which happens much more rarely.