Removed all log statements.
If I ever think it's worth it to have a logging mechanism that creates
log files, with like a dayly roll-over and a limited retention period,
I'll put the log statements back in. But at this point they are only
helpful during development. So I can just add them as needed and remove
them before committing.