Removed all priority key bindings.
This makes all modal dialogs truly modal, in that the user had to press
Esc first to get back the main screen, and kind quit the app directly
with Ctrl+Q from the dialog. I think that's fine. Because it can also
get in the way of a widget that lets the user press a key or key
combination to reassign to the binding/action.