Added key remapping to Settings dialog.
Implemented a new interactive key input widget that, when activated by
clicking or pressing Enter, records the key or key combination the user
presses and reports the new key to the parent widget by posting a
message.

Added a "Keys" panel to the Settings dialog that contains a key input
for every binding defined for the application or editor widget. The user
can now persistently change the key bindings to their liking.