Use Textual's internal key designations in configuration files.
We no longer allow a wide range of key notations in the configuration
file(s). Namely, before we would map "^Q" to "ctrl+q" or "Esc" to
"escape". But now we only accept the latter, i.e. the internal key
designations that Textual recognizes. (It ignores key designations that
it doesn't recognize.)

But since the Settings dialog now covers everything that can potentially
be configured in the files, and should continue to do so going forward
as more settings are added, we no longer expect users to edit the file
directly. Which is why we can be more strict about the expected syntax.
In other words, the content of the configuration file is considered an
an implementation detail and is subject to change. Though users may edit
the file manually at their own risk. And the location is openly
communicated (such as via `ked config files`) to allow easy backup and
transfer to other machines.

An added benefit is that our custom `KeyInput` widget is now independent
of other parts of the code base and could easily be dropped into any
other Textual application for reuse.