Renamed key binding for command palette from "Commands" to "Palette".
Initially, I liked the term "Commands" better, as it just lists all the
commands, while the term "Palette" could mean a lot of other things. But
since the default and most sensible key binding is ^P (as ^C is taken),
it's best to refer to it a "Palette" in the user interface so as to make
the key combination easier to remember.